Move the defaultListBoxMatchFilter in a new FilterUtils file
