+++ packages/react-vapor/src/components/select/SelectSelector.ts
@@ -4,7 +4,8 @@ import * as _ from 'underscore';
 import {IReactVaporState} from '../../ReactVapor';
 import {convertStringListToItemsBox} from '../../reusableState/customList/StringListReducers';
 import {DropSelectors} from '../drop/redux/DropReducers';
-import {FilterBoxSelectors, MatchFilter} from '../filterBox/FilterBoxSelectors';
+import {FilterBoxSelectors} from '../filterBox/FilterBoxSelectors';
+import {MatchFilter} from '../filterBox/FilterBoxUtils';
 import {IItemBoxProps} from '../itemBox/ItemBox';
 import {IListBoxState} from '../listBox/ListBoxReducers';
 import {ISelectWithFilterProps} from './hoc/SelectWithFilter';
