+++ packages/react-vapor/src/components/select/hoc/SelectWithFilter.tsx
@@ -12,7 +12,8 @@ import {UUID} from '../../../utils/UUID';
 import {Button, IButtonProps} from '../../button/Button';
 import {IFilterBoxOwnProps} from '../../filterBox/FilterBox';
 import {FilterBoxConnected} from '../../filterBox/FilterBoxConnected';
-import {FilterBoxSelectors, MatchFilter} from '../../filterBox/FilterBoxSelectors';
+import {FilterBoxSelectors} from '../../filterBox/FilterBoxSelectors';
+import {MatchFilter} from '../../filterBox/FilterBoxUtils';
 import {IItemBoxProps} from '../../itemBox/ItemBox';
 import {Svg} from '../../svg/Svg';
 import {ISelectOwnProps, ISelectSpecificProps} from '../SelectConnected';
